fix: address PostHog review feedback
