+++ jobs/tasks.py
@@ -27,6 +27,7 @@ from hn_jobs.posthog_events import (
     distinct_id_for_email,
     identify_company,
     model_from_feature_flag,
+    openai_usage_properties,
 )
 from hn_jobs.utils import get_tjalerts_logger
 from users.models import CustomUser
@@ -52,18 +53,6 @@ MAX_COMPANY_EMAILS_LENGTH = 2000
 REMOTE_OK_API_URL = "https://remoteok.com/api"
 REMOTE_OK_USER_AGENT = "gettjalerts.com jobs importer (https://gettjalerts.com)"
 MOJIBAKE_MARKERS = ("\u00c3", "\u00c2", "\u00e2", "\u00d8", "\u00d9")
-
-
-def openai_usage_properties(completion):
-    usage = getattr(completion, "usage", None)
-    if not usage:
-        return {}
-
-    return {
-        "prompt_tokens": getattr(usage, "prompt_tokens", None),
-        "completion_tokens": getattr(usage, "completion_tokens", None),
-        "total_tokens": getattr(usage, "total_tokens", None),
-    }
 
 
 def build_job_extraction_request(text):
